+++ tse_analytics/core/models/aggregation_combo_box_delegate.py
@@ -4,7 +4,7 @@ from tse_analytics.core.data.shared import Aggregation
 
 
 class AggregationComboBoxDelegate(QStyledItemDelegate):
-    aggregation_operations = [i for i in Aggregation]
+    aggregation_operations = (i for i in Aggregation)
 
     def paint(self, painter, option, index):
         if isinstance(self.parent(), QAbstractItemView):