+++ tse_analytics/views/settings/binning_settings_widget.py
@@ -1,11 +1,12 @@
 from PySide6.QtCore import QSize
 from PySide6.QtWidgets import QWidget
 
-from tse_analytics.core.data.binning import BinningMode, BinningParams
+from tse_analytics.core.data.binning import BinningMode
 from tse_analytics.core.manager import Manager
 from tse_analytics.core.messaging.messages import DatasetChangedMessage
 from tse_analytics.core.messaging.messenger import Messenger
 from tse_analytics.core.messaging.messenger_listener import MessengerListener
+from tse_analytics.modules.phenomaster.data.dataset import Dataset
 from tse_analytics.views.settings.binning_settings_widget_ui import Ui_BinningSettingsWidget
 
 
@@ -22,23 +23,29 @@ class BinningSettingsWidget(QWidget, MessengerListener):
 
         self.ui.applyBinningCheckBox.stateChanged.connect(self._apply_binning_changed)
 
-        self.ui.binningModeComboBox.addItems([e for e in BinningMode])
+        self.ui.binningModeComboBox.addItems((e for e in BinningMode))
         self.ui.binningModeComboBox.currentTextChanged.connect(self._binning_mode_changed)
+
+        self.dataset: Dataset | None = None
 
     def register_to_messenger(self, messenger: Messenger):
         messenger.subscribe(self, DatasetChangedMessage, self._on_dataset_changed)
 
     def _on_dataset_changed(self, message: DatasetChangedMessage):
-        self.ui.applyBinningCheckBox.setChecked(False)
-        self.ui.binningModeComboBox.setCurrentText(BinningMode.INTERVALS)
         if message.data is None:
+            self.dataset = None
+            self.ui.applyBinningCheckBox.setChecked(False)
+            self.ui.binningModeComboBox.setCurrentText(BinningMode.INTERVALS)
             self.ui.widgetTimePhasesSettings.clear()
         else:
-            self.ui.widgetTimeIntervalSettings.set_data(message.data.binning_settings.time_intervals_settings)
-            self.ui.widgetTimeCyclesSettings.set_data(message.data.binning_settings.time_cycles_settings)
-            self.ui.widgetTimePhasesSettings.set_data(message.data.binning_settings.time_phases_settings)
+            self.dataset = message.data
+            self.ui.widgetTimeIntervalSettings.set_data(self.dataset.binning_settings.time_intervals_settings)
+            self.ui.widgetTimeCyclesSettings.set_data(self.dataset.binning_settings.time_cycles_settings)
+            self.ui.widgetTimePhasesSettings.set_data(self.dataset.binning_settings.time_phases_settings)
+            self.ui.binningModeComboBox.setCurrentText(self.dataset.binning_settings.mode)
+            self.ui.applyBinningCheckBox.setChecked(self.dataset.binning_settings.apply)
 
-    def _binning_mode_changed(self, value: BinningMode):
+    def _binning_mode_changed(self, value: str):
         match value:
             case BinningMode.INTERVALS:
                 self.ui.widgetTimeIntervalSettings.setVisible(True)
@@ -52,19 +59,17 @@ class BinningSettingsWidget(QWidget, MessengerListener):
                 self.ui.widgetTimeIntervalSettings.setVisible(False)
                 self.ui.widgetTimeCyclesSettings.setVisible(False)
                 self.ui.widgetTimePhasesSettings.setVisible(True)
-        self._binning_params_changed()
+
+        if self.dataset is not None:
+            binning_settings = self.dataset.binning_settings
+            binning_settings.mode = BinningMode(self.ui.binningModeComboBox.currentText())
+            Manager.data.apply_binning(binning_settings)
 
     def _apply_binning_changed(self, value: int):
-        self._binning_params_changed()
-
-    def _binning_params_changed(self):
-        apply = self.ui.applyBinningCheckBox.isChecked()
-        mode = BinningMode(self.ui.binningModeComboBox.currentText()) if apply else BinningMode.INTERVALS
-        params = BinningParams(
-            apply,
-            mode,
-        )
-        Manager.data.apply_binning(params)
+        if self.dataset is not None:
+            binning_settings = self.dataset.binning_settings
+            binning_settings.apply = self.ui.applyBinningCheckBox.isChecked()
+            Manager.data.apply_binning(binning_settings)
 
     def minimumSizeHint(self):
         return QSize(300, 70)